net: wwan: Allow WWAN drivers to provide blocking tx and poll function
At the moment, the WWAN core provides wwan_port_txon/off() to implement
blocking writes. The tx() port operation should not block, instead
wwan_port_txon/off() should be called when the TX queue is full or has
free space again.

However, in some cases it is not straightforward to make use of that
functionality. For example, the RPMSG API used by rpmsg_wwan_ctrl.c
does not provide any way to be notified when the TX queue has space
again. Instead, it only provides the following operations:

  - rpmsg_send(): blocking write (wait until there is space)
  - rpmsg_trysend(): non-blocking write (return error if no space)
  - rpmsg_poll(): set poll flags depending on TX queue state

Generally that's totally sufficient for implementing a char device,
but it does not fit well to the currently provided WWAN port ops.

Most of the time, using the non-blocking rpmsg_trysend() in the
WWAN tx() port operation works just fine. However, with high-frequent
writes to the char device it is possible to trigger a situation
where this causes issues. For example, consider the following
(somewhat unrealistic) example:

 # dd if=/dev/zero bs=1000 of=/dev/wwan0p2QMI
 dd: error writing '/dev/wwan0p2QMI': Resource temporarily unavailable
 1+0 records out

This fails immediately after writing the first record. It's likely
only a matter of time until this triggers issues for some real application
(e.g. ModemManager sending a lot of large QMI packets).

One fairly simple solution for this is to allow WWAN drivers to provide
a potentially blocking tx() operation, and provide an additional poll()
operation to set additional poll flags. This integrates nicely with the
RPMSG API and does not break other WWAN drivers.

With these changes, the dd example above blocks instead of exiting
with an error.

Signed-off-by: Stephan Gerhold <stephan@gerhold.net>
---
I'm not sure if this is really a good idea but so far it's the only
simple solution I could think of. The main alternative would be
changing the RPMSG API (as well as likely all RPMSG drivers) so that
they can report/notify when the TX queue is full or has free space
again.

Alternative suggestions are very welcome, that's why this is a RFC. :)