iio: accel: bmc150: Drop duplicated documentation of supported chips
The chips supported by the bmc150-accel driver are clearly documented
in Kconfig, in the bmc150_accel_chip_info_tbl as well as in all the
device ID tables in the I2C/SPI drivers. It's easy to forget to update
the lists in the file header. Drop those entirely to reduce the amount
of changes required to add new chip variants.

Signed-off-by: Stephan Gerhold <stephan@gerhold.net>
---
New patch in v2. Originally I tried to reorder those too but then it
caused conflicts in all my following patches so I'm not convinced
it's worth to try and keep those up to date.