dt-bindings: arm: cpus: Document qcom,msm8916-smp enable-method
Since MSM8916 is an ARM v8 64-bit SoC the CPU cores are normally booted
using "psci" or "spin-table". Unfortunately, some devices come with signed
32-bit-only firmware without ARM64 and PSCI support. This setup is easy to
support since it's very similar to older 32-bit only Qualcomm SoCs
(e.g. MSM8226) and only requires adding a few new definitions to already
existing drivers.

Document the qcom,msm8916-smp enable-method. It is actually just an alias
of qcom,msm8226-smp since it should be implemented identically.
The enable-method is only valid on ARM32 and must not be used on ARM64.

Signed-off-by: Stephan Gerhold <stephan@gerhold.net>
---
Changes in v2:
  - Clarify that this is only supposed to be used on ARM 32-bit
    (Add comment to commit message and DT schema)

The "see above for ARM v8 64-bit" comment refers to this a few lines
above in the DT schema:

      # On ARM v8 64-bit this property is required
      - enum:
          - psci
          - spin-table