cpuidle: qcom-spm: Check if any CPU is managed by SPM
At the moment, the "qcom-spm-cpuidle" platform device is always created,
even if none of the CPUs is actually managed by the SPM. On non-qcom
platforms this will result in infinite probe-deferral due to the
failing qcom_scm_is_available() call.

To avoid this, look through the CPU DT nodes and check if there is
actually any CPU managed by a SPM (as indicated by the qcom,saw property).
It should also be available because e.g. MSM8916 has qcom,saw defined
but it's typically not enabled with ARM64/PSCI firmwares.

This is needed in preparation of a follow-up change that calls
qcom_scm_set_warm_boot_addr() a single time before registering any
cpuidle drivers. Otherwise this call might be made even on devices
that have this driver enabled but actually make use of PSCI.

Fixes: 60f3692b5f0b ("cpuidle: qcom_spm: Detach state machine from main SPM handling")
Reported-by: Marek Szyprowski <m.szyprowski@samsung.com>
Link: https://lore.kernel.org/r/86e3e09f-a8d7-3dff-3fc6-ddd7d30c5d78@samsung.com/
Signed-off-by: Stephan Gerhold <stephan@gerhold.net>
---
Daniel, would be great if you could ack this patch and PATCH 3/4
(the cpuidle part) if they look good to you. I think it's easiest if Bjorn
takes them together with the qcom_scm changes through the qcom tree.

Marek had an alternative fix for this [1], the difference in this patch is
that it avoids creating the platform device entirely if no CPU is managed
by a SPM.

[1]: https://lore.kernel.org/r/20211020120643.28231-1-m.szyprowski@samsung.com/