ASoC: dt-bindings: Add schema for "awinic,aw8738"
Add a DT schema for describing Awinic AW8738 audio amplifiers. They are
fairly simple and controlled using a single GPIO. The number of pulses
during power up selects one of a few pre-defined operation modes. This
can be used to configure the speaker-guard function (primarily the
power limit for the amplifier).

Signed-off-by: Stephan Gerhold <stephan@gerhold.net>
---
Changes in v2:
  - Rename "enable-gpios" -> "mode-gpios" as an attempt to better
    reflect its true purpose. Add description to prevent confusion.
  - Rewrite all descriptions to be more clear about the purpose of
    the amplifier operation modes.
  - Add maximum for "awinic,mode"